First draft at creating the multi-core version of Seeds
The process to create the feature simply bypasses the Node class, which
will start the JXTA network if used.  Instead, I modified the Worker
class, create DesktopWorker and NetworkWorker.  NetworkWorker does the
same thing as Worker used to do.  Worker is now the parent class for
DesktopWorker and NetworkWorker.  This to allow UserThread to receive an
instace of the generic Worker.  UserThread uses a map that is part of
the Worker class.

The next thing to modify was the MultiModePipe.  The classes (The Client
and Server) were modified to not require the Node object to be not null.
 Their code also enphasises the use of the shared memory connection over
the network.  This is the same as the initial implementation, but there
was some code rearrangement to make sure no method is called on a null
Node object.